pyproject.toml
[build-system]
requires = ["setuptools>=80", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "cheeky-shrimps"
version = "0.1.0"
description = "An AI-powered interactive knowledge graph and flashcard learning app"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "dash==4.0.0",
    "dash-bootstrap-components==2.0.4",
    "dash_cytoscape==1.0.2",
    "groq==1.1.1",
    "pypdf==6.9.1",
    "python-dotenv==1.2.2",
    "Flask==3.1.3",
    "plotly==6.6.0",
    "anyio==4.12.1",
    "blinker==1.9.0",
    "certifi==2026.2.25",
    "charset-normalizer==3.4.6",
    "click==8.3.1",
    "colorama==0.4.6",
    "distro==1.9.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.11",
    "itsdangerous==2.2.0",
    "Jinja2==3.1.6",
    "MarkupSafe==3.0.3",
    "narwhals==2.18.0",
    "nest-asyncio==1.6.0",
    "numpy==2.4.3",
    "packaging>=24.0",
    "pydantic==2.12.5",
    "pydantic_core==2.41.5",
    "pyparsing==3.3.2",
    "python-dateutil==2.9.0.post0",
    "retrying==1.4.2",
    "six==1.17.0",
    "sniffio==1.3.1",
    "tqdm==4.67.3",
    "typing_extensions==4.15.0",
    "urllib3==2.6.3",
    "Werkzeug==3.1.6",
]

[project.scripts]
cheeky-shrimps = "shrimps.app:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["shrimps*"]

[tool.setuptools.package-data]
shrimps = ["*.json", "*.md"]
